Wrong response type for this call
This call has a response that corresponds to ProfileDevicesLinkagesResponse, not ProfileResponse. According to the documentation it should be a CertificatesResponse (see https://developer.apple.com/documentation/appstoreconnectapi/list_all_certificates_in_a_profile), but that's not correct either, since the attributes are missing from the certificates in the data array.

An alternative option could be to make the attributes optional in certificate and use CertificatesResponse here.

For now I propose to abuse ProfileDevicesLinkagesResponse or else create a new response type specifically for this.

What do you think?